requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
python-mpd2==0.5.5
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.44.0
